Jenkins' job reliability and 8.0 support
As this job is used in jenkins as non-stop job
(i.e. it should be running until aborted), there is a possibility
that executor thread will receive more than one interrupt,
and will generate SIGTERM during the cleanup procedure.
This commit adds some signal ignorance to the cleanup
procedure to have a chance of successful completion if
multiple signals are arriving.

Also, the work on 8.0+ support is started.

Change-Id: If78b896b15d87f560d7d26d4da5becfc757658ef