Role-based and individual updates
Added functionality to make role-based and particular nodes updates,
i.e. update only computes or ceph-osds. This is needed to divide
update process into stages which is usefull on very big environments.
This could be also useful to rerun the update process on a particular
node. Also the time interval variable is added. Using this variable
it is possible to update some senstive part of a cluster giving a
time to synchronize its states.

Bumped the header and the version.

Change-Id: If638107ec6d4180b4006f7565742687b6f20f2bb